sched-analyzer.bpf.c: Proper way to read comm
Not exactly sure I understood it, but we can't just get a pointer and
hope for the best. We should read the whole array of chars.

Getting a pointer to the string is not the correct way to access them.
We should at least use bpf_read_kernel_str() to read it from kernel
memory and copy it into our bpf program where we can use it as we like.

Not sure how we could read it, maybe becaue bpf_printk() executes from
kernel context so it could read the string pointer. I can't explain why
this pointer gets cleared if we don't immediately call bpf_printk()
after reading it.

Signed-off-by: Qais Yousef <qyousef@layalina.io>