Makefile: produce a static binary
So that we can use it on any system, especially Android, without having
to have the right shared binaries available.

Size of binary increased by 1MB from 1.4MB to 2.4MB.

We should make this optional if this ever to be distributed as a binary.
Static libraries are recipe for security problems. For
personal/development, static is a good start.

Signed-off-by: Qais Yousef <qyousef@layalina.io>