sched-top: Fix a crash when showing the rq
We could end up reading partial row at the end and end up with NA
values, which causes plotext to crash. Drop NA values.

Also increase refresh rate to 5 seconds as the python script is quite
heavy. As this matures we'll provide an option for the user to select
the refresh rate. For now, let's see what it takes to get it to do what
we want it to do.

Signed-off-by: Qais Yousef <qyousef@layalina.io>