sched-top: Better management of last_ts
We assumed sched-analyzer is started by sched-top before and we always
start from time 0. But now sched-analyzer starts separately, this
assumption is broken as it could have been running for a while.

This is more robust way to look at the last N seconds of the captured
trace regardless of how long it has been running for.

Signed-off-by: Qais Yousef <qyousef@layalina.io>