Remove unnecessary sleep()
ring_buffer__poll() essentially uses epoll(7) to wait for the data, which plays
similar act as sleep() when there has no event.  Not to mention we've
already have a 1000ms timeout set in ring_buffer__poll().

Remove the sleep() to prevent potential lose of the event, which
exacerbates when the machine is busy.