sched-analyze: move each event to its own pthread
This should allow us to process each event in parallel and avoid the
potential to drop events because of the sequential nature of the current
design where we poll each buffer one after the other. With the new
design we should poll each buffer in parallel and ensure they get
processed faster.

This will not eliminate the problem of a buffer overflowing problem
entirely. But if it happens, it will be due to the buffer size being too
small for the current rate of events. We could try to increase the
buffer size then.

Signed-off-by: Qais Yousef <qyousef@layalina.io>