sched-analyzer: Re-introduce the sleep in the poll loop
Effectively reverts ca6c20d ("Remove unnecessary sleep()").

Now we run all events in threads, the busy ones end up in an busy loop
as we never hit the timeout.

Force a sleep, but rather a short 10ms so that these threads don't end
up being a busy loop.

The timeout on ring_buffer__poll() is only how long to sleep if nothing
is found. But the new usleep() is to ensure we go idle between polls.

Signed-off-by: Qais Yousef <qyousef@layalina.io>