android: Makefile: Clone elfutils which is required by android
Android NDK only has x86 binaries to create a build; but we need to
cross compile to arm64. It is hard to get this cross compile deps done
easily with distros and NDK doesn't contain everything like header files
for elfutils. So clone it to get what we need from it reliably.

Signed-off-by: Qais Yousef <qyousef@layalina.io>