Make vmlinux.h part of the build
If a structure changed, we could fail to compile the code as we still
need to have a header file with these items defined correctly.

BPF provides a mechanism to check if the field actually exists at
runtime by using bpf_core_field_exists() call which will do a runtime
check. But for compile time we need a header that contain these fields
still. So make vmlinux.h part of the tree to ensure it'll always build.

Signed-off-by: Qais Yousef <qyousef@layalina.io>