perfetto_wrapper: Fix track uniqueness issues
We have a unique setup where different events are generated by the
kernel belong to the same thread/process which perfetto uses to create
unique track IDs. This causes TRACE_EVENT() to overlap and
terminute/nest in a weird way.

Implement a TRACK_ID() wrapper to create a unique number for each event
we generate to handle the parallel nature of the events generated.

Signed-off-by: Qais Yousef <qyousef@layalina.io>