sched-analyzer-pp/idle: Show TUI sequentially
showing subplots can mess up subsequents plots as to clear the graph we
must call clf() but doing so will clear the theme.

Stick to the simpler formula of showing each plot sequentially which is
fine for TUI. Maybe we can change the saved output from matplotlib in
a similar fashion, but leaving it as-is for now.

Signed-off-by: Qais Yousef <qyousef@layalina.io>