sched-analyzer-pp/sched: Sort-by max and 90%
Each gives us different info. Worst case scenarios for max, and how good
are we doing on 'average' with 90% sorting. We want both numbers to be
low. Tasks with highest max might be outliers. Tasks with highest 90%
demonstrate reasonable consistency in being in a certain state (runnable
being the most indicative of latencies).

Signed-off-by: Qais Yousef <qyousef@layalina.io>