sched-analyzer-pp/sched: Use simple_bar instead of simple_stacked_bar
The latter prints a nice concise graph, but can't be copied into text
emails as it requires colours to be readable, which is nice on a local
terminal, but not for text based emails.

Also the simple graph gives easier to read break down of time spent in
state/CPU.

Add a printed text about the Sum Time in State (ms) similar to Sum Time
Running on CPU (ms). It displays percentage of time beside the sum and
would aid the visual graph to see how much a task spent in states for
the duration of its life. The absolute sum values are still more
meaningful as the percentage could be high, but the task could have
barely spent any time running which would be visible from look at
percentage alone.

Signed-off-by: Qais Yousef <qyousef@layalina.io>