sched-analyzer-pp/sched: Show % time as graphs instead of text
The graphs look busy now, so not 100% about this conversion, but it is
seems more informative this way. The Sum tells us in absolute time how
long the task ran for. But % gives us a good view on residency in each
state.

There could be better way to divide the output to help with readability.

Signed-off-by: Qais Yousef <qyousef@layalina.io>