sched-analyzer-pp/freq: Fix issue when freq is constant till end of trace
If the frequency is constant till the end of the trace, then we won't
have any frequency updates till the end of the trace and we'll
mistakenly display part of the frequencies omitting the whole constant
period at the end of the trace.

This usually happens when looking at use cases where the system was busy
and running at max speed for a long time till the end of the trace.

Signed-off-by: Qais Yousef <qyousef@layalina.io>