sched-analyzer-pp/freq: Add option to plot freq seen by a task
With some data manipulation, we should be able to combine info from freq
and sched states to plot the cpu frequency seen by the task when it is
RUNNING. Would be useful to analyze the perceived performance by the
task.

We also provide option to show the freq residency for the frequencies
seen by the task when it is RUNNING.

This still needs more testing and verification.

Signed-off-by: Qais Yousef <qyousef@layalina.io>