sched-analyzer-pp/freq: Better handle timestamp convertion and alignment
Freq events are not emitted frequently enough and we can have long gaps
with no events. This includes at start or end. The latter was handled in
commit 6de4985 ("sched-analyzer-pp/freq: Fix issue when freq is constant
till end of trace") but the same problem can happen due to missing data
from the beginning of the trace. This makes the timestamp of the plot
inaccurate due to the missing data.

Reorganize the code to better handle the timestamp conversion and
upsampling. Instead of doing it at the df_freq level, move it to act on
df_freq_cpu level as this produces more accurate upsampling results with
correct time stamps of trace_start/end injected to this isolated
dataframe. Attempting to fix it at df_freq level requires more work.
Defer instead when we filter for a particular CPU.

Clean up some dead code to calculate duration when plotting frequency.

Signed-off-by: Qais Yousef <qyousef@layalina.io>