sched-analyzer-pp: Add new utils module
It's a better place to encapsulate trace_start/end_ts variables and for
the generic convert_ts function and other potentially common code.

Convert all modules to use the new utils.convert_ts() function.

For sched module do the conversion for a track like we need to do that
for each CPU freq. Doing the conversion globally produces wrong results
as the dataframe is a combination of multiple tracks.

Signed-off-by: Qais Yousef <qyousef@layalina.io>