sched-analyzer-pp: Make --freq-task more accurate
We should account only for the RUNNING time. Previously we were
filtering for RUNNING time but this loses information when we switch to
another state and ended up accounting for other states.

Properly omit the sections where the task is not RUNNING.

Increase the reindexing rate to 100us instead of 1ms to provide more
accurate results. Higher sampling rate will slow things down
considerably.

Signed-off-by: Qais Yousef <qyousef@layalina.io>