sched-analyzer-pp: Multip sa_track with df_tid_running
For tracks that show per-task values, show the values when the task is
running by default.

Generalize the logic from freq module to show per-task freq to be
applied to other modules easily.

Initialize df_states in utils module as it is now a common data that is
required by several modules and provide a function for other modules to
use to get it.

Signed-off-by: Qais Yousef <qyousef@layalina.io>