sched-analyzer-pp/utils: Fix duration accuracy
We need to dropna() to get accurate total_duration, but with absolute
mode now we can see that the residency calculation after multiplication
is not correct compared to the sum running time from sched-states
because we dropna().

Only dropna() when calculating total_duration instead for the whole
dataframe. abs and % residencies look more accurate now.

Signed-off-by: Qais Yousef <qyousef@layalina.io>